requirements.txt
streamlit
sentence-transformers
anthropic
numpy
python-dotenv
torch
